pyproject.toml
[tool.poetry]
name = "strong-password-generator"
version = "0.1.0"
description = ""
authors = ["Daniel Vergara <daniel.vergara.echeverri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.4"
SQLAlchemy = "^1.4.31"
python-dotenv = "^0.19.2"
SQLAlchemy-Utils = "^0.38.2"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = "strong_password_generator.main:generate"
history = "strong_password_generator.main:history"
